% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsgd.R
\name{dsgd_single}
\alias{dsgd_single}
\title{Disease-Specific Gene Detection (Wrapper Function)
Fitting the model with the option of single or multiple correlated tissues and the option of adding interaction terms.}
\usage{
dsgd_single(
  list_y,
  matrix_x,
  method = "VI",
  niter = 2000,
  nwarmup = 1000,
  nchain = 3
)
}
\arguments{
\item{list_y}{a list of binary variables to indicate the spot status}

\item{matrix_x}{a matrix of covariates, with the coordinates placed in the final two columns of the matrix}

\item{interaction}{optional, character vector specifying the interaction terms}
}
\value{
a fitted stan mode
}
\description{
Disease-Specific Gene Detection (Wrapper Function)
Fitting the model with the option of single or multiple correlated tissues and the option of adding interaction terms.
}
\examples{
\donttest{}
}
