% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_imputation.R
\name{missing_imputation}
\alias{missing_imputation}
\title{Title}
\usage{
missing_imputation(list_y, coords)
}
\arguments{
\item{list_y}{a list of binary variables to indicate the spot status}

\item{coords}{a matrix with two columns to indicate the coordinate information for each spot/cell}
}
\value{
imputed outcome dataset
}
\description{
Title
}
\examples{
\donttest{}
}
