% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_plot.R
\name{plot_coef}
\alias{plot_coef}
\title{Plot top beta coefficients}
\usage{
plot_coef(fit, matrix_x, n = 30)
}
\arguments{
\item{fit}{A stanfit object}

\item{matrix_x}{a matrix of covariates, with the coordinates placed in the final two columns of the matrix}

\item{n}{number of coefficients to be plotted}
}
\value{
a ggplot object
}
\description{
Plot top beta coefficients
}
\examples{
\donttest{}
}
