% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_plot.R
\name{plot_network}
\alias{plot_network}
\title{Network plot}
\usage{
plot_network(model, interaction, label_size = 13, node_color = "#C1D8E7")
}
\arguments{
\item{model}{A fitted \code{stan} model object returned from \code{dsgd()}.}

\item{interaction}{A vector specifying covariate names for interaction terms.}

\item{label_size}{Size of the node labels in the plot.}

\item{node_color}{Color of the nodes in the network plot.}
}
\value{
A ggplot object showing the interaction network.
}
\description{
Network plot
}
\examples{
\donttest{}
}
