% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict}
\alias{predict}
\title{Prediction}
\usage{
predict(fit, data, sweep = 100)
}
\arguments{
\item{fit}{A fitted \code{stan} model object returned from \code{dsgd()}.}

\item{data}{A dataframe of covariates and coordinates. The last two columns must be the spatial coordinates.}

\item{sweep}{The number of Gibbs sampling iterations.}
}
\value{
A vector of predicted binary outcomes.
}
\description{
Prediction
}
\examples{
\donttest{}
}
